pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
authors = [
  { email = "30631553+liblaf@users.noreply.github.com", name = "liblaf" },
]
dependencies = [
  "confz>=2.0.1,<3",
  "dvclive>=3.48.0,<4",
  "fast-simplification>=0.1.9,<0.2",
  "jaxtyping>=0.2.34,<0.3",
  "jinja2>=3.1.4,<4",
  "lazy-loader>=0.4,<0.5",
  "loguru>=0.7.2,<0.8",
  "matplotlib>=3.9.2,<4",
  "meshio[all]>=5.3.5,<6",
  "numpy>=2.0.2,<2.1.0",
  "open3d>=0.18.0,<0.19",
  "pandas>=2.2.3,<3",
  "polars>=1.9.0,<2",
  "pooch>=1.8.2,<2",
  "prettytable>=3.11.0,<4",
  "pydantic-settings>=2.5.2,<3",
  "pydicom>=3.0.0,<4",
  "pytetwild>=0.1.dev1,<0.2",
  "pyvista>=0.44.1,<0.45",
  "rich>=13.9.2,<14",
  "ruamel-yaml>=0.18.6,<0.19",
  "scipy>=1.14.1,<2",
  "seaborn>=0.13.2,<0.14",
  "sortedcollections>=2.1.0,<3",
  "sparse>=0.15.4,<0.16",
  "taichi>=1.7.2,<2",
  "tetgen>=0.6.5,<0.7",
  "tomlkit>=0.13.2,<0.14",
  "trimesh[all]>=4.4.7,<5",
  "ubelt>=1.3.6,<2",
]
description = "Add a short description here"
name = "mesh-kit"
requires-python = ">=3.11,<3.12"
version = "0.0.1"

[project.optional-dependencies]
cpu = ["jax>=0.4.31,<0.5", "pytorch3d"]
cu121 = ["jax[cuda12]>=0.4.31,<0.5"]
dev = [
  "dvc>=3.54.1,<4",
  "icecream>=2.1.3,<3",
  "marimo>=0.9,<0.10",
  "pdoc>=15,<16",
  "pyright>=1.1.377,<2",
  "ruff>=0.6.2,<0.7",
]
docs = [
  "black>=24.8.0,<25",
  "mkdocs-gen-files>=0.5.0,<0.6",
  "mkdocs-git-committers-plugin-2>=2.3.0,<3",
  "mkdocs-git-revision-date-localized-plugin>=1.2.7,<2",
  "mkdocs-literate-nav>=0.6.1,<0.7",
  "mkdocs-material[imaging]>=9.3.1,<10",
  "mkdocs-section-index>=0.3.9,<0.4",
  "mkdocs>=1.6.1,<2",
  "mkdocstrings-python>=1.11.1,<2",
  "mkdocstrings>=0.26.1,<0.27",
]
test = ["pyperf>=2.7.0,<3", "pytest-benchmark>=4.0.0,<5", "pytest>=8.3.2,<9"]

[tool.hatch.build.targets.wheel]
packages = ["src/mkit"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pixi.dependencies]
numpy = ">=2.0.2,<2.1.0"

[tool.pixi.environments]
cpu = ["cpu", "dev", "docs", "test"]
cu121 = ["cu121", "dev", "docs", "test"]
default = ["cu121", "dev", "docs", "test"]

[tool.pixi.feature.cu121]
channels = ["nvidia"]

[tool.pixi.feature.cu121.dependencies]
pytorch = { build = "*_cuda12.1_*", channel = "pytorch", version = ">=2.4.1,<3" }
pytorch-cuda = { channel = "pytorch", version = ">=12.1,<13.0" }
pytorch3d = { build = "*_cu121_*", channel = "pytorch3d", version = ">=0.7.8,<0.8" }
torchvision = { channel = "pytorch", version = ">=0.19.1,<0.20" }

[tool.pixi.project]
channel-priority = "disabled"
channels = ["anaconda", "conda-forge", "pytorch", "pytorch3d"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
mesh-kit = { editable = true, path = "." }

[tool.pixi.pypi-options]
extra-index-urls = ["https://miropsota.github.io/torch_packages_builder"]

[tool.pixi.system-requirements]
libc = { family = "glibc", version = "2.35" }
linux = "6.5"

[tool.pixi.tasks]
docs-build = "mkdocs build"
docs-deploy = "mkdocs gh-deploy --force --no-history"
docs-serve = "mkdocs serve"
gen-init = "scripts/gen-init.sh"
test = "pytest"
